feat: adding smtp features

+++ Devocean.Tests/IntegrationTestBase.cs
@@ -1,6 +1,9 @@
+using System.Reflection;
 using AutoMapper;
 using Devocean.Core.Application.Interfaces;
+using Devocean.Core.Application.Mappers.Common;
 using KellermanSoftware.CompareNetObjects;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
@@ -9,13 +12,13 @@ using Xunit.Abstractions;
 
 namespace Devocean.Tests;
 
-public abstract class IntegrationTestBase<TStartup, TDbContext> : IDisposable 
+public abstract class IntegrationTestBase<TStartup, TDbContext> : IDisposable
     where TStartup : class
     where TDbContext : DbContext, IDbContext
 {
     private readonly bool _shouldEnsureDeleted;
     private readonly bool _shouldEnsureCreated;
-    private readonly bool _shouldMidrate;
+    private readonly bool _shouldMigrate;
     static readonly object _locker = new();
 
     protected CompareLogic _compareLogic;
@@ -25,22 +28,40 @@ public abstract class IntegrationTestBase<TStartup, TDbContext> : IDisposable
     protected WebApplicationFactory<TStartup> Factory { get; set; }
     protected HttpClient Client { get; set; }
 
-    protected IntegrationTestBase(ITestOutputHelper testOutputHelper, Action<DbContextOptionsBuilder>? optionsAction,
-        Action<TDbContext> seedFunc, bool shouldEnsureDeleted = false, bool shouldEnsureCreated = false, bool shouldMidrate = false)
+    protected IntegrationTestBase(CompareLogic compareLogic, ITestOutputHelper testOutputHelper, TDbContext dbContext,
+        IMapper mapper, WebApplicationFactory<TStartup> factory, HttpClient client)
+    {
+        _compareLogic = compareLogic;
+        TestOutputHelper = testOutputHelper;
+        DbContext = dbContext;
+        Mapper = mapper;
+        Factory = factory;
+        Client = client;
+    }
+
+    protected IntegrationTestBase(ITestOutputHelper testOutputHelper,
+        Func<IWebHostBuilder, Action<DbContextOptionsBuilder>> dbContextOptionsBuilder,
+        bool shouldEnsureDeleted = false, bool shouldEnsureCreated = false,
+        bool shouldMigrate = false,
+        Action<TDbContext>? seedFunc = null,
+        Assembly? automapperProfileAssembly = null)
     {
         _shouldEnsureDeleted = shouldEnsureDeleted;
         _shouldEnsureCreated = shouldEnsureCreated;
-        _shouldMidrate = shouldMidrate;
+        _shouldMigrate = shouldMigrate;
         TestOutputHelper = testOutputHelper;
         _compareLogic = new CompareLogic();
+        
+        AutomapperProfile.IncludedAssemblies.Add(typeof(Program).Assembly);
+        if (automapperProfileAssembly != null) AutomapperProfile.IncludedAssemblies.Add(automapperProfileAssembly);
 
         Factory = new WebApplicationFactory<TStartup>()
             .WithWebHostBuilder(builder =>
             {
                 builder.ConfigureServices(services =>
                 {
-                    services.AddAutoMapper(typeof(TStartup)); 
-                    services.AddDbContext<TDbContext>(optionsAction);
+                    services.AddAutoMapper(typeof(AutomapperProfile).Assembly);
+                    services.AddDbContext<TDbContext>(dbContextOptionsBuilder(builder));
                 });
                 // ... Configure test services
             });
@@ -55,10 +76,10 @@ public abstract class IntegrationTestBase<TStartup, TDbContext> : IDisposable
 
         lock (_locker)
         {
-            DbContext.Database.EnsureDeleted();
-            DbContext.Database.EnsureCreated();
-            DbContext.Database.Migrate();
-            seedFunc(DbContext);
+            if (_shouldEnsureDeleted) DbContext.Database.EnsureDeleted();
+            if (_shouldEnsureCreated) DbContext.Database.EnsureCreated();
+            if (_shouldMigrate) DbContext.Database.Migrate();
+            seedFunc?.Invoke(DbContext);
         }
     }
